Patch nerf
Goof was supposed to do this in his pr but he never did. He approves of this change though.

Patches now only hold 40u. This is 1:1 with goon and makes them less of a direct upgrade from pills with a tiny delay.

Medkit patches now only hold 20u by default. As they are now, patches will heal you to full health with like one or two. That means a medkit is more useful than a doctor. This fixes that and is also 1:1 with goon.