AI hostile lockdown power now re-opens doors
Fixes #18985.

A side effect is that when the mass door opening happens, it's not
instant like it was before. I did consider making a global proc to call
to do that instead, but it made more sense to have the timer attached to
the doors themselves, rather than the AI or a global proc.