Minor projectile refactor
Haven't tested shit, I'll download the zip and try this out in a bit.

Basically we account for tick drifts or byond not firing this proc on time such that we run it quicker if it falls behind.

This also allows for any speed, even when it doesn't round to world.tick_lag, causing it to just alternate rounding such that it avgs to the requested speed.

Something better would be to move all of this to a subsystem, to reduce the amount of sleeps and wakeups, but i don't think there will ever be enough projectiles to justify the effort of doing so.