Small devil fixes (#19302)
* Removes errant cast from devilinfo.add_soul()

* Clarifies a banishment ritual in the Codex Gigas.

Sprinkling holy water isn't enough, it needs to be in the devil's bloodstream.

* Infernal jaunt now correctly extinguishes the user.
Revival contracts no longer actually set the target on fire, but rather uses a fake fire.

* Pitchforks don't burn devils or soulless on use/pickup anymore.
Adds cast charge time to infernal jaunt, so that it's impossible to jaunt while already jaunting.
Adds feedback to using infernal jaunt to jaunt in.