Makes black box code less retarded
Not returning when qdeling for dupe (from killing a second colossus) caused byond to ***hard delete*** the entire contents, as they wouldn't be in the contents when destroy() ran, but would be when the garbage collector hard deleted it.

Hard deleting the 200 some items people decided to put in this thing takes awhile, a long ass while.

Also made the code faster and less loopy and more up-to-date with current systems.