Tweaks space wind probability curve.
Basically, before move_prob would be 50% if the pressure was exactly equal to the atoms resistance value, 100% if it was exactly double, etc.

Now that number is 75% if the difference is exactly pressure_resistance, but we minus 25% from the number, meaning it's still 50% if the difference is exactly pressure_resistance, but it raises and lowers fasters otherwise.

We also don't even attempt to move if the move_prob is lower than 25%, to avoid the issue where even small differences make everything move over time.