Fixes lava being retarded
By looping thru the turf on entered, it caused more issues then it needed to. By making it only act on the new thing on entered, and loop thru the turf only on process(). This avoids all sorts of issues and fixes it applying a burn multiple times quickly to everything if a bunch of things entered the turf at once.