Makes transit space unjauntible, and removes a few spells restriction from the wizard den. There's nothing there you can't reach normally now except a mob that wants to kill you and some meat. (#20853)
Fixes an area error that was allowing jaunting from the wizard den after a LOT of creativity.
The radios in the wizards den should work now.
Fixes a rouge airless tile in the den.
Recompiles z2 as a dmm as it was erroring on mapmerge and it was in a very strange form that didn't match up with other dmm formated maps. It looks like everything is ruined, but I assure you everything just swapped around.