Hardlight Bows (#20944)
Adds hardlight bows and hardlight quivers.
Does 12 brute and 18 burn on hit.
The quivers regenerate one arrow every few seconds and is worn on the back.
Arrows disintegrate on hit and are destroyed after 20 seconds if it is dropped to prevent spam.

ADDED TO UPLINK AT 7 TELECRYSTALS FOR A BOW AND QUIVER. Can be included in surplus crates.
:cl: