Changes astar from using a turf var to a proc list (#21947)
Fixes a runtime related to two bots trying to path through the same turf, and should improve the overall cost.

The runtime was caused by one bot pathing through a square, then getting check tick sleeped and the second both pathing through the same square.

Because the pnode lived on the turf it would assume that it had already seen that turf and try to resort the open heap, causing a bad index runtime as that turf was not in it's own open list