[s]fixes ai turf click admin spam (#21957)
* fixes ai turf click spam

get_turf_pixel returns null if handed anything that's not a movable. I could istype it before I call that, but this looks cooler and it's about the same result since get_turf_pixel already istypes it.

* Does proper checks.

Clicks on static count as a click on the turf it's covering, and ais are supposed to be allowed to click on static to center their screen on that turf. so have to only count clicks on objects/mobs.