EMPs cause chameleon clothing to flicker (#22086)
* [DNM] EMPs cause chameleon clothing to flicker

:cl: coiax
add: Chameleon clothing produced by the syndicate has been found to
react negatively to EMPs, randomly switching forms for a time.
/:cl:

- EMPing them when they're in the box is fine, but when they're being
worn, the game slows down a LOT, which I assume is the constant spamming
of the regenerate_icons() proc on the mob every time the chameleon
clothes change, which if they're wearing all 10 of them, that's a lot of
redundant calls.

* EMPed camoclothing no longer causes lag