Tomb of the Dragoon, and the Sky Bulge lance (#22270)
* Sky Bulge Commit mk2

WORK THIS TIME.

* Indentation.

* Jordbord's suggestions

Incorporated. The bone spear line was also doubletabbed I was set up!

* FUCKING INDENTATION

* the flan update

they don't exist in world but here have a mob

* we fireflans now

also isliving is in now, just gotta actually map a thing

* the tomb update

it's bigger now

* .gitignore revert

* workpls

* Revert ".gitignore revert"

This reverts commit 1fefd34fc7c0dbc8f1f353a95e53665e03ba75cf.

* workthistime

* Revert "workthistime"

This reverts commit 214509d5abda941c470708431536ccd715334e8e.

* well it's the closest gitignore fix so far

* fix dme change

will need to be loaded for the map part to work though

* gitignorerevert take 43

* jordbord fix mk 2

* includes flan.dm in the .dme