Simplifies some wall code. (#22281)
To put it simply, this adds the vars girder_type and sheet_amount to the wall code, and cuts down on a small amount of now unneeded lines.

This simplifies it so adding newer walls and different wall/girder combinations is slightly simpler. It is also easier to adjust how much sheets a wall drops when deconstructed and what type of girder is left behind.

Tested and everything works as intended. Please inform me if I mixed anything.