Throttles preferences hrefs
The user will not be allowed to trigger a click on a preferences href until their previous one has finished.

rshoe is using it spam us.

I know somebody was talking about trying to do a more proper fix, such as by making it not re-generate these when it's the same, so i'll test merge this until 24 hours or they make such a pr.