VROOM (#23311)
🆑 BASILMAN YOUR MAIN MAN
add: Added BM SPEEDWAGON THE BEST (AND ONLY) SPACE CAR ON THE MARKET.
/🆑
I PRESENT THE BM SPEEDWAGON

THE FRONTIER IN VEHICLE TECHNOLOGY, DRIVE ON THE PURE POWER OF THE STARS, RUN PEOPLE OVER, MOVE REALLY FAST DOWN THE HALL, PICK UP HOES*

this is intended for wizards at some point but ill leave that for another pr because i fear the balance meme flocks.

this was fun to make, and hopefully fun to watch operate.

eliminates #23170