Transformation diseases do not change your name (#23499)
:cl: coiax
add: A victim of a transformation disease will retain their name.
tweak: The slime transformation disease can turn you into any colour or
age of slime.
/:cl:

Polymorphed people keep their name, so why not transformation diseases.
And colourful slimes are fun, and the slime disease requires a high tier
of slime anyway, and if you're going to infect humans with the disease
in order to harvest slimes, then that's probably a feature.