Makes Icon Smoothing faster (#23689)
* Makes Icon Smoothing faster

During world init or large explosions, doing an associated add on this list accounted for a good chunk of the overhead.

I removed the check for if the subsystem existed because subsystems are created before atoms or the map is created.

In another branch of the code, this removed 5 seconds (out of 60) from world+SS init

* Fix edge case.