Engraved messages UI and fixes (#23700)
* Engraved messages UI

:cl: coiax
add: Engraved messages now have a UI, which any player, living or dead
can access. See when the message was engraved, and like or dislike
accordingly.
add: Admins have additional options with the UI, seeing the player ckey,
original character name, and the ability to outright delete messages at
the press of a button.
/:cl:

- If we were to move to a system where higher rated messages would be
more likely to be selected, this provides the ability to get those
ratings.
- Ratings are by ckey, so character independent.
- Authors cannot upvote or downvote their own messages.
- Did I mention that admins can just now delete the message, instead of
mucking around varediting them?

* Arrows > thumbs

* Brings back colours

* Readds message erasure

* Revert "Readds message erasure"

This reverts commit 92ef8e3f1672c0afd4d6cd630c329d6bf4689e77.

* Revert "Brings back colours"

This reverts commit 7a784e4e463398342c6ee5e18dc885f2a7cd46d0.