[s]More security exploit patching
Thanks to Nadrew at byond forums

The full details is that somebody can use the built in version of these verbs to aimbot or find the location of items.

.click 'name-of-mob' and you will click on that mob, meaning you now point in that direction and if its in range, you click directly on it. something that could be automated with auto hotkey.

This still requires a control_freak bypass, but there is a way to do that that is stupid easy that I don't want to detail it.