Posters refactor (#24548)
* Posters refactor

- Posters now use a simple type-per-poster definition, rather than some
large global lists requiring sprite numbers to be the same as entries in
a list.
- Putting a specific poster on a map now simply means putting the
specific type that you want, rather than var editing (often
unsuccessfully).
- There are random all-posters, random contraband posters, and random
official posters for mappers too.
- Ripped posters are now their own type.
- The potential to make additional classes of posters has been left
open, I'd like to move a bunch of the advertisement posters to their own
catagory, and clean up the "official" category to be more obviously
on-brand.

* Updates DeltaStation

* Changed the maps of Metastation

* OmegaStation posters updated

* Updated Pubby posters

* Map merged?

* Updated away mission and ruin posters

* Removes vars from garbage