Deathrattles follow links jump to turf if no mob is found (#24797)
* Deathrattles also give a turf link

:cl: coiax
add: When someone dies, there will be a (T) link to jump to the turf
where they died, on the offchance they died in a manner that left no
body.
/:cl:

Supermatter disintegrations, singularity eatings, other stuff that
doesn't leave a body means that the (F) link doesn't work. But the turf
should still be there, so click on that instead.

* Jump to mob if exists, turf if not