Stacks automatically merge on Initialize (#24810)
* Stacks automatically merge on Initialize

:cl: coiax
fix: Stacks of materials will automatically merge together when created.
You may notice differences when ejecting metal, glass or using the cash
machine in the vault.
/:cl:

Okay, so instead of just obsessively checking every time we make a
stackable item to see if anything stacks (which not everything does
anyway), we put the check in Initialize. Obviously, you can't merge into
a full stack, so it just goes down the line.

Objects are allowed to qdel() in New/Initialize, we made that a thing
that was possible.

Also some generic code cleanup, mostly returns into `.`

* Don't reuse variable names