Generates messages when xeno eggs are delivered (#24852)
* Generates messages when xeno eggs are delivered

:cl: coiax
add: The egg spawner on Metastation will generate a station message and
inform the admins if an egg is spawned. (It's only a two percent chance,
but live in hope.)
/:cl:

Metastation has a 2% egg spawner. If it picks the xeno egg, it does so
silently, telling no one. Now it informs the station and also admins and
the game log.

To reiterate, this behaviour is already in the map, I'm just generating
logging and messages. If admins want to simulate this, they can spawn a
`/obj/effect/spawner/xeno_egg_delivery` in any location they like.

* Xeno delivery makes a command report instead

I also refactored command report to avoid the duplication of the SECRET
COMMAND REPORT announcement.