TGUI GPS (#26561)
Ported the GPS system to TGUI.
GPS signals may now be updated automatically or manually, based on a new toggle.
The GPS now has the option to scan only the same Z-level, or all of them.
Toggling GPS power may now be done from the interface as well as the alt-click.
The user's own GPS unit is no longer displayed in the signal list, and is now its own section for quick and easy self-location.
Signals on the same z-level now have distance and direction information added to aid with hunting and exploration.
The scanning range and auto-update buttons have tooltips to explain how they work.