Plasma vessels regenerate slowly off resin (#26778)
:cl: coiax
balance: Plasma vessels, the organs inside aliens that are responsible
for their internal plasma storage, will regenerate small amounts of
plasma even if the owner is not on resin weeds.
/:cl:

- Rate is 10% of normal. Stops aliens accidentally spending their plasma
poorly, and then being unable to plant weeds ever again. Breathing in
plasma is not really an option for most aliens, and not many people know
about it.
- This is an "anti-stupidity" feature, in essence.