Ports dakimakuras from HippieStation. (#27167)
* Ports dakimakuras from HippieStation.

Dakimakuras are life sized body pillows with the image of anime
characters on them. This was originally an 8ch creation, but was ported
to Hippie. I'm now porting it here.

Although I am permabanned from /tg/, I know that /tg/ has the most
talented, creative, and active coders of all code bases. If I add
Dakimakura's here, then you guys will undoubtedly add new content to it.
I will then be able to enjoy that new content on other servers that
mirror your pull requests. Makes sense yes?

* Removes Hotwheels from list of selectable dakis.

Hotwheels was the host of 8ch and I removed the image from the DMI. This
would cause a runtime error if not removed.

* Fixes what AnturK told me to.

* Uses to_chat.

* Most of those changes done.

* Apply's PR fixes