Unprotects the deadmin list
The only thing being in this list does is give you a re-admin verb that still verifies you are really suppose to be an admin by checking the .txt, and make you exempt for bans and a few other on connection things.

An admin adding somebody to this list with vv is not a credible enough of a security threat to justify locking it, and I needed to edit this list to test something, and i couldn't, and that made me sad.