Changes clickcatcher to dynamically scale based on view range (#29370)
Scale()s the icon when range is below 16, if it's above 16, scales the icon to 16 and transforms the rest of the way.

This replaces the other method of creating one click catcher image per tile in the range of the users view, which was pretty memory hungry and inefficient. A single icon operation while the view range changes should still be pretty performant and require less memory bookkeeping.l

This ensures it works when the users view range changes, but does not result in a loss of precision for view ranges up to 16 when moving the mouse (it triggers per pixel which simply get larger when transformed)

fixes #29342