Makes bIcon scale to 16x16 like byond chat output did (#29575)
* Makes bIcon scale to 16x16 like chat output did

There was lag and ddos issues because to_chat does certain find replace operations on the resulting string, so putting base64 encoded data of large images in these strings would be insanely and excessively costly.

This is basically banking on the assumption that md5 on a 480x480 image is less costly than 4 find/replaces and 2 html_encodes.

If this doesn't work or we need larger icon support for some reason, I'll just move bIcon over to the asset cache system.