Adds high-contrast inhand templates for spriters (#29716)
The new templates are in __inhand_template.dmi which is in the base inhands folder.

I made this because it's often unclear what parts of inhands should be cut out to accomodate for the actual "hand" portion. Very bright colors signify something that should block a sprite (as if it's being held behind that part) and shaded parts are something that the sprite should be held over. For instance, the north-facing sprite has bright colors for all parts (as the human is in the way of the object) but the south-facing one is dark for all parts because it's being held directly towards the camera.

There are also more generic "solid" sprites, for those that don't care about the parts.