Adds callback select system
This system lets you run a group of callbacks asynchronously, returning once all of them finish, without having to wait for each callback to finish before running the next one.

I made it in my bicon pr, but i don't need it anymore, so i removed it from there and put it here.