ssgarbage now runs every 2 seconds.
This was moved down to 5ds back when subsystems had to complete their entire workload each fire in the old mc system.

The idea was that less would have gotten in the queue to be hard deleted in 5ds compared to 20ds, and a tiny amount of lag more often would be more preferred to alot of lag less often.

That is no longer the case now.