Make add_details faster
Storing all of these intermediary strings in the string table might be unbalancing it (especially since there would be a high number of partially matching strings.

Waiting until later to concat the string will prevent this.

add_details was getting to be more expensive then atmos, this is very like why (but I can't be sure)