Gives engineering cyborgs an item they can use in place of circuits for basic construction (#30464)
Adds the electroadaptive pseudocircuit - technobabbly name intended - to engineering cyborgs. They can use this in place of circuit boards during construction of alarms, APCs, and firelocks to create a circuit at the cost of varying amounts of power. There's a brief cooldown between uses to prevent borgs from becoming circuit farms on wheels.

Note that in order to create circuits, you need to cannibalize simple circuits (anything the pseudocircuit can make) to get the material. Each pseudocircuit starts with five circuits by default.

It can also be used on a completed APC frame with no power cell to create an empty, low-capacity power cell in the APC at the cost of 500W of power; this is to allow borgs to repair a broken APC without external aid because the CE won't bring the fucking cell over get over here you brainless fuck.

Code-wise, this just adds extra checks to the frames' attackby procs that do the same thing as the other circuits. Airlock electronics, computer boards, and machine circuits cannot be made using this, and it's only for basic construction purposes.