Stops airlocks from getting harddeleted, also fixes an unrelated runtime (#31412)
* Stops airlocks from getting harddeleted, also fixes an unrelated runtime

Destroying doors with a sniper rifle or anything that flat out destroys them good would runtime in doors.dm #L184.
Also doors would very often not get qdel'd properly, it was not 100% of the time but very common, this *seems* to fix it but since reproducing the bug was never quite certain, I'm not absolutely sure I got it. I did destroy all the doors on the station without triggering a harddel on a door after this fix, where previously it would happen.

* old friend ex_act