Fixes excessive overlay churn  (#32065)
* Fixes excessive overlay chern

Adding an overlay on every cross and uncross of every player (with how bloody the station gets this is called on just about every fucking movement, some times multiple times if there are multiple types of stains (oil, blood, alien, etc) is an excessive as fuck amount of chern in the overlay subsystem.

This is likely the cause of overlays being so high.

* I APOLOGIZE FOR NOTHING!